+++ test/regressions/test_lake_creator.py
@@ -241,7 +241,7 @@ def test_lakecreator_extractor(zip_file, kube: TestClient):
     # Logic to wait till OrbitJob creates
     lakecreator.wait_until_ready(timeout=60)
     # Logic to pass or fail the pytest
-    lakecreator.wait_until_job_completes(timeout=1200)
+    lakecreator.wait_until_job_completes(timeout=7200)
     current_status = lakecreator.get_status().get("orbitJobOperator").get("jobStatus")
     logger.info(f"current_status={current_status}")
     # Cleanup
@@ -336,7 +336,7 @@ def test_lakecreator_glue_table_creator(datafile, kube: TestClient):
     # Logic to wait till OrbitJob creates
     lakecreator.wait_until_ready(timeout=60)
     # Logic to pass or fail the pytest
-    lakecreator.wait_until_job_completes(timeout=1200)
+    lakecreator.wait_until_job_completes(timeout=7200)
     current_status = lakecreator.get_status().get("orbitJobOperator").get("jobStatus")
     logger.info(f"current_status={current_status}")
     # Cleanup
@@ -398,7 +398,7 @@ def test_lakecreator_lf(kube: TestClient):
     # Logic to wait till OrbitJob creates
     lakecreator.wait_until_ready(timeout=60)
     # Logic to pass or fail the pytest
-    lakecreator.wait_until_job_completes(timeout=1200)
+    lakecreator.wait_until_job_completes(timeout=7200)
     current_status = lakecreator.get_status().get("orbitJobOperator").get("jobStatus")
     logger.info(f"current_status={current_status}")
     # Cleanup